java

+++ app/src/main/java/com/example/myapplication/RecruitActivity.java
@@ -23,12 +23,10 @@ import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.firestore.SetOptions;
+import com.google.firebase.firestore.QueryDocumentSnapshot;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.Objects;
 
 
@@ -38,14 +36,15 @@ public class RecruitActivity extends AppCompatActivity {
     private EditText title_et; // 제목 에디터텍스트
     private Spinner people_sp; // 총인원 스피너
     private Spinner place_sp; //place 스피너
-    private TextView date_view1; //날짜 텍스트
-    private TextView date_view2; //시간 텍스트
+    private TextView date_view; //날짜 텍스트
+    private TextView time; //시간 텍스트
     private RadioButton radio_together; //같이 라디오버튼
     private RadioButton radio_solo; //따로 라디오버튼
     private final FirebaseFirestore db = FirebaseFirestore.getInstance(); // 파이어 베이스 store
     private Context rContext; //context
     private FirebaseAuth firebaseAuth;     //파이어베이스 인증처리
     private final Board board = new Board(); // 게시판
+    private String id, place_name, showtitle, showpeople, showinfo, dateview;
     long mNow;
     Date mDate;
     @SuppressLint("SimpleDateFormat")
@@ -68,18 +67,26 @@ public class RecruitActivity extends AppCompatActivity {
         radio_together = findViewById(R.id. radio_together);
         //따로 라디오버튼
         radio_solo = findViewById(R.id.radio_solo);
-        //다른 인텐트 받아올때 사용
-        Intent anotherintent = getIntent();
         //현재날짜
-        date_view1 = findViewById(R.id.date_view1);
+        date_view = findViewById(R.id.date_view1);
         //날짜 지정
-        date_view1.setText(getTime());
+        date_view.setText(getTime());
         //시간 ID 가져오기
-        date_view2 = findViewById(R.id.date_view2);
+        time = findViewById(R.id.date_view2);
+
+        //다른 인텐트 받아올때 사용
+        Intent intent = getIntent();
+        place_name = intent.getStringExtra("place_name");
+        showtitle = intent.getStringExtra("title_info");
+        showpeople = intent.getStringExtra("people_number");
+        showinfo = intent.getStringExtra("delevery_info");
+        dateview = intent.getStringExtra("time");
+        id = intent.getStringExtra("id");
 
         //시간 데이터 가져오기
-        String data = anotherintent.getStringExtra("time");
-        date_view2.setText(data);
+        String data = intent.getStringExtra("time");
+        time.setText(data);
+
         // 총인원 스피너
         people_sp = findViewById(R.id.people_sp);
         people_sp.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
@@ -98,40 +105,59 @@ public class RecruitActivity extends AppCompatActivity {
     }
     // 삭제하기 버튼 이벤트
     public void deletebtn(){
-        /*db.collection(place_name).document(board_lists.get(position).getDoc_id())
-                .delete()
-                .addOnSuccessListener(new OnSuccessListener<Void>() {
-                    @Override
-                    public void onSuccess(Void aVoid) {
-                        Log.d(TAG, "DocumentSnapshot successfully deleted!" );
+        db.collection("post") //컬렉션 게시물 선택
+                .get() // 데이터 가져옴
+                .addOnCompleteListener(task -> {
+                    if (task.isSuccessful()) {
+                        for (QueryDocumentSnapshot document : task.getResult()) {
+                            Board Board = document.toObject(Board.class); // 오브젝트 형식으로 변환.
+                            FirebaseUser user = firebaseAuth.getCurrentUser();
+                            if(user.getUid().equals(Board.getUid())){ // 로그인 되어있는 회원의 uid와 게시판에 저장된 uid와 일치할 경우 삭제
+                                db.collection("post")
+                                        .document(id)//doc id와 일치할 경우 삭제
+                                        .delete()
+                                        .addOnSuccessListener(aVoid -> Toast.makeText(rContext, " 정상적으로 게시글이 삭제되었습니다. ", Toast.LENGTH_SHORT).show())
+                                        .addOnFailureListener(e -> Toast.makeText(rContext, " 게시글을 찾지 못했습니다. ", Toast.LENGTH_SHORT).show());
+                            }else{
+                                Toast.makeText(rContext, " 삭제할 수 있는 권한이 없습니다. ", Toast.LENGTH_SHORT).show();
+                            }
+                        }
                     }
-                })
-                .addOnFailureListener(new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception e) {
-                        Log.w(TAG, "Error deleting document", e);
-                    }
-                });*/
+                });
     }
     // 수정하기 버튼 이벤트
     public void retouchbtn(){
-        /*DocumentReference washingtonRef = db.collection("cities").document("DC");
-
-        // Set the "isCapital" field of the city 'DC'
-        washingtonRef
-                .update("capital", true)
-                .addOnSuccessListener(new OnSuccessListener<Void>() {
-                    @Override
-                    public void onSuccess(Void aVoid) {
-                        Log.d(TAG, "DocumentSnapshot successfully updated!");
+        db.collection("post") //컬렉션 게시물 선택
+                .get() // 데이터 가져옴
+                .addOnCompleteListener(task -> {
+                    if (task.isSuccessful()) {
+                        for (QueryDocumentSnapshot document : task.getResult()) {
+                            Board Board = document.toObject(Board.class); // 오브젝트 형식으로 변환.
+                            FirebaseUser user = firebaseAuth.getCurrentUser();
+                            if(user.getUid().equals(Board.getUid())){ // 로그인 되어있는 회원의 uid와 게시판에 저장된 uid와 일치할 경우 삭제
+                                if (radio_together.isChecked()){
+                                    board.setTogether(radio_together.getText().toString()); // 같이하면 같이 저장
+                                }else if(radio_solo.isChecked()){
+                                    board.setSolo(radio_solo.getText().toString()); // 따로하면 따로 저장
+                                }
+                                db.collection("post")
+                                        .document(id)//doc id와 일치할 경우 수정
+                                        .update("board_name",title_et.getText().toString(),
+                                                "time",time.getText().toString(),
+                                                "delevery_info",delevey_et.getText().toString(),
+                                                "solo",radio_solo.getText().toString(),
+                                                "together",radio_together.getText().toString(),
+                                                "people_number",people_sp.getSelectedItem().toString()
+                                                )
+                                        .addOnSuccessListener(aVoid -> Toast.makeText(rContext, " 정상적으로 게시글이 삭제되었습니다. ", Toast.LENGTH_SHORT).show())
+                                        .addOnFailureListener(e -> Toast.makeText(rContext, " 게시글을 찾지 못했습니다. ", Toast.LENGTH_SHORT).show());
+                            }
+                            else{
+                                Toast.makeText(rContext, " 수정할 수 있는 권한이 없습니다. ", Toast.LENGTH_SHORT).show();
+                            }
+                        }
                     }
-                })
-                .addOnFailureListener(new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception e) {
-                        Log.w(TAG, "Error updating document", e);
-                    }
-                });*/
+                });
     }
 
     // 등록하기 버튼 함수
@@ -146,6 +172,7 @@ public class RecruitActivity extends AppCompatActivity {
         startActivity(placeBoardintent);//화면전환
         finish();//등록게시판 종료
     }
+
     // 시간선택 버튼 기능 = 타임피커
     public void rider_time(View view) {
         Intent intent = new Intent(RecruitActivity.this, TimePickerActivity.class); //타임피커 activity에 대한 화면전환 객체
@@ -153,12 +180,11 @@ public class RecruitActivity extends AppCompatActivity {
     };
     // 함수로 데이터 정보 받아옴
     private void save_boarddata(Board board){
-        Map<String, Object> data = new HashMap<>(); // 문서 id 자동 생성
         FirebaseUser user = firebaseAuth.getCurrentUser();
         board.setBoard_name(place_sp.getSelectedItem().toString()); // 장소 스피너 텍스트 저장
         board.setPeople_number(people_sp.getSelectedItem().toString()); //인원수 스피너 텍스트 저장
-        board.setDate(date_view1.getText().toString()); //날짜 저장
-        board.setTime(date_view2.getText().toString()); //시간 저장
+        board.setDate(date_view.getText().toString()); //날짜 저장
+        board.setTime(time.getText().toString()); //시간 저장
         board.setDelevery_info(delevey_et.getText().toString()); //배달내용 저장
         board.setTitle_info(title_et.getText().toString()); //제목 저장
         assert user != null;
@@ -203,7 +229,7 @@ public class RecruitActivity extends AppCompatActivity {
                  if(result.getResultCode()==RESULT_OK){
                      //타임피커의 액티비티 값을 텍스트뷰에 가져옴
                      assert result.getData() != null;
-                     date_view2.setText(result.getData().getStringExtra("time"));
+                     time.setText(result.getData().getStringExtra("time"));
                  }
             }
     );
